refactor(workspace): move BrainConfig + IdleConfig into claudectl-core
App.rs holds typed `BrainConfig`, `IdleConfig`, `HealthThresholds` fields.
HealthThresholds already lives in `claudectl-core::health`; the brain and
idle structs were still defined in the binary's `src/config.rs`, which
means moving `app.rs` to a future TUI crate (#275) would have dragged
binary dependencies along for type signatures alone.

Carved both structs out:

* New `claudectl-core::config` module: `BrainConfig`, `IdleConfig`,
  `IdleTask`, `default_test_runners()`. Pure data — no parsing, no
  AgentConfig dep, nothing that couples to binary subsystems. The
  binary still owns TOML parsing and CLI flag layering.
* `src/config.rs` re-exports them so existing `crate::config::BrainConfig`
  call sites in main.rs, brain/client.rs, brain/engine.rs, brain/evals.rs,
  and app.rs resolve unchanged.

That removes the last typed binary-crate refs from app.rs's struct
definition (besides the still-binary brain/coord/hive/relay subsystem
calls, which are next).

Verified: `cargo build --all-features`, `cargo test --all-features`,
`cargo clippy --all-features -- -D warnings`, `cargo fmt --check`, plus
`cargo test -p claudectl-core --lib` (115 passed) all clean.

Toward #275 (extract claudectl-tui crate).
